extract renderHistogram()
